requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pydantic
pydantic-settings
python-multipart
httpx
pandas
openpyxl